Move TracingSystemSetting to interceptor pkg
